+++ opentelemetry-exporter-google-cloud/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add mapping between opentelemetry and google traces attributes
+  ([#90](https://github.com/GoogleCloudPlatform/opentelemetry-operations-python/pull/90))
+
 ## Version 0.15b0
 
 Released 2020-11-04